+++ src/server/server_game.h
@@ -31,11 +31,11 @@ typedef duration<float> dur_f;
 class Game : public Thread {
 private:
   std::mutex m;
-  std::map<int,std::shared_ptr<Queue<GameState>>> queues_sender;
+  std::map<uint8_t,std::shared_ptr<Queue<GameState>>> queues_sender;
   Broadcaster broadcaster;
   Queue<std::shared_ptr<RunnableCommandGame>> commands;
   uint8_t game_id;
-  int last_player_id_added = 0;
+  uint8_t last_player_id_added = 0;
   bool keep_playing = true;
   bool started = false;
   GameManager game_manager;
@@ -47,9 +47,9 @@ public:
   explicit Game(uint8_t &game_id);
 
   Queue<std::shared_ptr<RunnableCommandGame>>* add_player(std::shared_ptr<Queue<GameState>> sender_queue, 
-											  const int& player_id);
+											  const uint8_t& player_id);
 
-  void delete_player(const int &player_id);
+  void delete_player(const uint8_t &player_id);
 
   void handle_command();
 
@@ -67,7 +67,7 @@ public:
 
   bool is_dead();
   
-  int get_game_id();	//Provisional
+  uint8_t get_game_id();	//Provisional
 
   Game(const Game &) = delete;
   Game &operator=(const Game &) = delete;
